+++ core/src/ten_manager/designer_frontend/src/flow/node/ExtensionNode.tsx
@@ -6,24 +6,34 @@
 //
 
 import {
-  type Node,
+  type Connection,
+  type Edge,
   type NodeProps,
   Position,
   useReactFlow,
 } from "@xyflow/react";
-import { BlocksIcon, SaveOffIcon } from "lucide-react";
+import {
+  AudioLinesIcon,
+  BlocksIcon,
+  DatabaseIcon,
+  SaveOffIcon,
+  TerminalIcon,
+  VideoIcon,
+} from "lucide-react";
 import React, { useCallback } from "react";
+import { useTranslation } from "react-i18next";
 import { Badge } from "@/components/ui/Badge";
 import { Button } from "@/components/ui/Button";
 import { DropdownMenuItem } from "@/components/ui/DropdownMenu";
+import { BaseHandle } from "@/components/ui/react-flow/BaseHandle";
 import { BaseNode } from "@/components/ui/react-flow/BaseNode";
-import { LabeledHandle } from "@/components/ui/react-flow/LabeledHandle";
 import {
   NodeHeader,
   NodeHeaderActions,
   NodeHeaderMenuAction,
   NodeHeaderTitle,
 } from "@/components/ui/react-flow/NodeHeader";
+import { Separator } from "@/components/ui/Separator";
 import {
   Tooltip,
   TooltipContent,
@@ -32,43 +42,70 @@ import {
 } from "@/components/ui/Tooltip";
 import { cn } from "@/lib/utils";
 import type { IExtensionNodeData, TExtensionNode } from "@/types/flow";
+import { EConnectionType } from "@/types/graphs";
+import { dispatchCustomNodeActionPopup } from "@/utils/events";
 
-export function ExtensionNode({ id, data }: NodeProps<TExtensionNode>) {
-  const { updateNodeData, setNodes } = useReactFlow();
+export function ExtensionNode({
+  id,
+  data,
+  isConnectable,
+}: NodeProps<TExtensionNode>) {
+  const { setNodes } = useReactFlow();
+
+  const { t } = useTranslation();
 
   const handleDelete = useCallback(() => {
     setNodes((nodes) => nodes.filter((node) => node.id !== id));
   }, [id, setNodes]);
+
+  const isInstalled = React.useMemo(() => {
+    return data.is_installed;
+  }, [data.is_installed]);
 
   return (
     <BaseNode className="h-fit p-0">
       <div className="-top-7 absolute right-2">
         <Badge>
           <BlocksIcon className="me-1 size-3" />
-          Ext node
+          {t("node.extension.tag")}
         </Badge>
       </div>
       <NodeHeader
         className={cn(
           "bg-gray-200 dark:bg-gray-800",
           "rounded-t-sm px-2 py-1",
-          "text-secondary-foreground"
+          "min-w-sm",
+          "text-secondary-foreground",
+          {
+            "text-ten-icontext-2": !isInstalled,
+          }
         )}
       >
-        <TooltipProvider delayDuration={50}>
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <Button variant="ghost" size="xs" className="cursor-help">
-                <SaveOffIcon className="size-4" />
-              </Button>
-            </TooltipTrigger>
-            <TooltipContent>
-              <p>Uninstalled</p>
-            </TooltipContent>
-          </Tooltip>
-        </TooltipProvider>
-
-        <NodeHeaderTitle>Mock title with a long name</NodeHeaderTitle>
+        {!isInstalled && (
+          <TooltipProvider delayDuration={50}>
+            <Tooltip>
+              <TooltipTrigger asChild>
+                <Button variant="ghost" size="xs" className="cursor-help">
+                  <SaveOffIcon className="size-4" />
+                </Button>
+              </TooltipTrigger>
+              <TooltipContent>
+                <p>{t("node.uninstalled")}</p>
+              </TooltipContent>
+            </Tooltip>
+          </TooltipProvider>
+        )}
+        <div className="font-roboto">
+          <NodeHeaderTitle className="text-lg">{data.name}</NodeHeaderTitle>
+          <div
+            className={cn(
+              "text-ten-icontext-2",
+              "font-light text-sm leading-none"
+            )}
+          >
+            {data.addon}
+          </div>
+        </div>
         <NodeHeaderActions>
           <NodeHeaderMenuAction
             className="cursor-pointer"
@@ -79,10 +116,15 @@ export function ExtensionNode({ id, data }: NodeProps<TExtensionNode>) {
         </NodeHeaderActions>
       </NodeHeader>
 
-      <div className="my-10 flex items-center gap-2">content here</div>
+      {/* <div className="my-10 flex items-center gap-2">content here</div> */}
 
-      <footer className={cn("w-full bg-gray-100", "rounded-b-sm")}>
-        <LabeledHandle
+      <footer
+        className={cn(
+          "flex w-full flex-col font-roboto text-xs",
+          "rounded-b-sm"
+        )}
+      >
+        {/* <LabeledHandle
           title="CMD"
           type="source"
           position={Position.Right}
@@ -93,8 +135,180 @@ export function ExtensionNode({ id, data }: NodeProps<TExtensionNode>) {
           type="source"
           position={Position.Right}
           className="mr-1"
+        /> */}
+        <Separator className="w-full" />
+        <HandleGroupItem
+          data={data}
+          isConnectable={isConnectable}
+          // onConnect={onConnect}
+          connectionType={EConnectionType.CMD}
+        />
+        <Separator className="w-full" />
+        <HandleGroupItem
+          data={data}
+          isConnectable={isConnectable}
+          // onConnect={onConnect}
+          connectionType={EConnectionType.DATA}
+        />
+        <Separator className="w-full" />
+        <HandleGroupItem
+          data={data}
+          isConnectable={isConnectable}
+          // onConnect={onConnect}
+          connectionType={EConnectionType.AUDIO_FRAME}
+        />
+        <Separator className="w-full" />
+        <HandleGroupItem
+          data={data}
+          isConnectable={isConnectable}
+          // onConnect={onConnect}
+          connectionType={EConnectionType.VIDEO_FRAME}
         />
       </footer>
     </BaseNode>
   );
 }
+
+const HandleGroupItem = (props: {
+  data: IExtensionNodeData;
+  isConnectable: boolean;
+  onConnect?: (params: Connection | Edge) => void;
+  connectionType: EConnectionType;
+}) => {
+  const { data, isConnectable, onConnect, connectionType } = props;
+
+  const handleClickDetails =
+    ({
+      type,
+      source,
+      target,
+    }: {
+      type?: EConnectionType;
+      source?: boolean;
+      target?: boolean;
+    }) =>
+    () => {
+      dispatchCustomNodeActionPopup({
+        action: "connections",
+        source: data.name,
+        target: undefined,
+        metadata: {
+          filters: {
+            type,
+            source,
+            target,
+          },
+        },
+      });
+    };
+
+  return (
+    <div className="my-0.5 flex items-center justify-between gap-x-4">
+      <div className="flex items-center gap-x-2">
+        <BaseHandle
+          key={`target-${data.name}-${connectionType}`}
+          type="target"
+          position={Position.Left}
+          id={`target-${data.name}-${connectionType}`}
+          // label={data.name}
+          // labelOffsetX={0}
+          isConnectable={isConnectable}
+          onConnect={onConnect}
+        />
+        <ConnectionCount
+          onClick={handleClickDetails({
+            type: connectionType,
+            target: true,
+          })}
+        >
+          {connectionType === EConnectionType.CMD && (
+            <span>{data.src[connectionType]?.length || 0}</span>
+          )}
+          {connectionType === EConnectionType.DATA && (
+            <span>{data.src[connectionType]?.length || 0}</span>
+          )}
+          {connectionType === EConnectionType.AUDIO_FRAME && (
+            <span>{data.src[connectionType]?.length || 0}</span>
+          )}
+          {connectionType === EConnectionType.VIDEO_FRAME && (
+            <span>{data.src[connectionType]?.length || 0}</span>
+          )}
+        </ConnectionCount>
+      </div>
+      <Button
+        size="sm"
+        variant="ghost"
+        className={cn("flex items-center gap-x-1", {
+          "cursor-pointer": handleClickDetails,
+        })}
+        onClick={handleClickDetails({
+          type: connectionType,
+        })}
+      >
+        {connectionType === EConnectionType.CMD && (
+          <TerminalIcon className="size-3 shrink-0" />
+        )}
+        {connectionType === EConnectionType.DATA && (
+          <DatabaseIcon className="size-3 shrink-0" />
+        )}
+        {connectionType === EConnectionType.AUDIO_FRAME && (
+          <AudioLinesIcon className="size-3 shrink-0" />
+        )}
+        {connectionType === EConnectionType.VIDEO_FRAME && (
+          <VideoIcon className="size-3 shrink-0" />
+        )}
+        <span>{connectionType.toUpperCase()}</span>
+      </Button>
+      <div className="flex items-center gap-x-2">
+        <ConnectionCount
+          onClick={handleClickDetails({
+            type: connectionType,
+            source: true,
+          })}
+        >
+          {connectionType === EConnectionType.CMD && (
+            <span>{data.target[connectionType]?.length || 0}</span>
+          )}
+          {connectionType === EConnectionType.DATA && (
+            <span>{data.target[connectionType]?.length || 0}</span>
+          )}
+          {connectionType === EConnectionType.AUDIO_FRAME && (
+            <span>{data.target[connectionType]?.length || 0}</span>
+          )}
+          {connectionType === EConnectionType.VIDEO_FRAME && (
+            <span>{data.target[connectionType]?.length || 0}</span>
+          )}
+        </ConnectionCount>
+        <BaseHandle
+          key={`source-${data.name}-${connectionType}`}
+          type="source"
+          position={Position.Right}
+          id={`source-${data.name}-${connectionType}`}
+          // label={data.name}
+          // labelOffsetX={0}
+          isConnectable={isConnectable}
+        />
+      </div>
+    </div>
+  );
+};
+
+const ConnectionCount = (props: {
+  children: React.ReactNode;
+  onClick?: () => void;
+}) => {
+  const { children, onClick } = props;
+
+  return (
+    <Button
+      size="sm"
+      variant={"ghost"}
+      className={cn("w-8 rounded-md bg-muted px-1 py-0.5 text-center text-xs", {
+        "cursor-pointer": onClick,
+      })}
+      onClick={onClick}
+    >
+      {children}
+    </Button>
+  );
+};
